Fix SBT ambiguous keys error reported for crossVersion
Remove the crossVersion setting since SBT will automatically set it to
`CrossVersion.binary` when `crossPaths` is `true` which is the default.
As for the cause, defining it inside `inThisPlugin` creates a conflict
with SBT's default which is automatically added to each subproject.

Fixes the following error reported for `show crossVersion`:

    akka-http-root > show crossVersion
    [error] Ambiguous keys: {.}/*:crossVersion, *:crossVersion
    [error] show crossVersion
    [error]                  ^