=htc #616 replace OutputTruncationException for the http client use case
We will now send an `UnexpectedConnectionClosureException` which more
closely explains the problem at hand.

Also, the error is reported with `failStage` instead of throwing which
prevents the noisy logging.

Removed `extends NoStackTrace` from the affected exceptions to allow users
to figure out where the exception was thrown (even if the stack trace
will always be the same).

Fixes #616.