Comment for Migration guide
Migrating from Spray to Akka-http, the thing that is most missing in my opinion is the ability to complete a request outside of the router. 
This guide (https://markatta.com/codemonkey/blog/2016/08/03/actor-per-request-with-akka-http/) helped me out a lot - especially the last section "Complete that context". As you've told me, this is intentional. But it still needs to be explicitly explained in the migration guide.

Some more documentation that was missing for me was a clear explanation about the differences between the bindAndHandleXXX versions.

Thank you very very much!