Use media type parameters in content negotiation (#963)
https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.1 describes how
to interpret the Accept header during content negotiation.

Notably, it specifies "The media type quality factor associated with a given
type is determined by finding the media range with the highest precedence which
matches that type", and then gives some examples.

'matching' is implemented in akka-http in `MediaRange::matches`, but for
`MediaRange.One` fails to take into account parameters.

'precedence' is implemented in akka-http when determining the sorting of
`acceptedMediaRanges` in `MediaTypeNegotiator`, but for non-wildcard media
types sorts by the qValue instead of by the number of parameters.