Add extra test to narrow down failure #443
I agree with Johannes that the only code path to this failure is when the
parboiled dispatcher would throw `RuleNotFoundException`, but don't see how
this could be nondeterministic. To further narrow this down I added a test that
fails specifically in that case, should be interesting to see if that one also
fails (and whether they fail together).