[backport from master] =htc #1724 automatically close connection when slot state -> Unconnected
Likely the bug happened in the Idle state. That's the one state that
didn't explicitly close the connection when the signal came. On the
other hand, the assumption may have been that on reception of a
`onConnectionCompleted` / `onConnectionFailed` signal, a connection was
already thought to be 'closed'.

This will also make the code simpler because SlotContext.closeConnection()
doesn't need to be exposed to the state handlers.

In some way it also shows the redundancy of the whole state handling code
structure as the code doesn't do much more than showing what expected signals
are in which states. The only side-effects left for the states are pushing
requests and results. But even those are never delayed so there's little logic
left in SlotStates.