Various cleanups all around
The remaining test failure looks legitimate: we've limited the number of
places that are allowed to create new streams in this PR, and we now no
longer supports setting a window size before registering the stream. Can
be fixed I guess, but we'd have to see if we consider it worth it.