=htc NewHostConnectionPool: remove weird pushRequestToConnectionAndThen HACK
The effect of the previous solution was that we had recursion of `updateState`
which we otherwise try to avoid really hard. It had a real effect on logs
where you could see the recursion in weird nesting of "Before" and "After"
log messages.

The new solution is to use another "virtual" state (+ event) to avoid the side-effecting
call in SlotState but instead add another state post processing clause.