Set a more descriptive message to the RejectionError exception
The toString implementation inherited by java.lang.Throwable only prints
the class name followed by the result of the getMessage method (which,
by default, returns null). As a consequence, the rejection carried by a
RejectionError instance is not shown.

We pass the given rejection to the RuntimeException constructor so that
it is shown as the exception message.