Remove test dependency on akka-typed
Because akka-typed is not available for 2.11, and we still want to test
akka-http on 2.11.

In the future we might want to re-introduce a dependency on akka-typed. At that
point we should decide whether to drop support for scala 2.11 from akka-http
entirely, or put in place some machinery to be able to exclude the parts that
depend on typed when testing with 2.11.