Use companion for custom header names
When the HeaderMagnet is formed for a user-defined custom header, it
used the class' name to populate the `headerName` field. However, the
user may have named their class differently than its name, by excluding
dashes, for example.

This could produce an incosistency in behavior when routes would
generate a `MissingHeaderRejection` for a custom header because the
`headerValueByType` directive constructs a `MissingHeaderRejection` with
a `headerName` populated from the magnet, which in the case of a custom
header, comes from the header's class name rather than its name field.

Since the magnet implicits have access to the user-defined header name
in the `ModeledCustomHeaderCompanion` object, we should take advantage
of it: This change updates the implicit to override the `headerName`
behavior by setting it to the companion's `name` field.

I considered using an alternate approach of analyzing the class passed
into the `ModeledCompanion.nameFromClass` method. However, that required
using reflection and potentially had broader side-effects since it would
change the underlying behavior of a public trait.