Accept RST on an already-half-closed HTTP/2 stream #2957
When grpccurl is used on a Akka gRPC server that does not support server
reflection, we send a 404, and grpcurl sends a RST with error code
PROTOCOL_ERROR.

We definitely didn't handle the RST correctly in that scenario. This PR
fixes handling the RST.

Separate from this PR, we should also investigate why grpcurl is sending
a PROTOCOL_ERROR: this indicates a HTTP/2 protocol-level error, while us
sending a 404 is an application-level error. It might be that grpcurl is
just incorrectly sending a PROTOCOL_ERROR here, or that we're in fact doing
something else wrong that triggers it.

The test is not very nice, but I don't see another way.