core: add server-side streamCancellationDelay to mitigate cancellation race conditions
Refs #2116

In the process, HttpServerSpec was changed back to use a regular
scheduler since cancellation dependent code now needs the scheduler
to make progress towards connection termination.

In the best case, we wouldn't be so reliant on cancellation for
proper termination / cleanup of streams. In reality, it seems we
still are, especially in this test.

Tests that explicitly test timers were moved into HttpServerWithExplicitSchedulerSpec.