core: prepend headers in `HttpMessage.addHeader` (#3029)
* core: prepend headers in `HttpMessage.addHeader`

Since by default headers are stored in a list, we would slowly append to the
list. It's questionable whether List is the right datastructure here but
that can be changed separately if needed.

Fixes #3028

* core: avoid `mapHeaders` inside of HttpMessage implementations

Since `mapHeaders` is a higher-order function it has a higher chance of
preventing inlining. Inside of HttpMessage implementations, mapHeaders is
not even simpler than using `withHeaders` directly.