Reuse UriParser In HttpRequestParser (#3108)
* parsing: allow reusing of parser including ValueStack

This is not quite done API-wise. You can now reuse a parser by providing
different values for `input`. Allocating the ValueStack turned up a lot
in profiling.

* core: precalculate size for ByteStringParserInput

* core: reuse UriParser in HttpRequestParser

Together with the other changes, this seems to be a big performance
improvement. Mostly, because the ValueStack is now only allocated once
and then reused across requests.