test: in WithLogCapturing check for and mark severe log messages
The idea is that tests should prove that we are "log-clean",
i.e. no warnings or errors are emitted accidentally.

For now the feature has to enabled by overriding `failOnSevereMessages`
for a whole test suite.

If warnings or errors are expected they should be asserted on and muted
by using log filters.

In some cases, we try to handle erroneous conditions gracefully but log
a warning in which case we might miss this abnormal condition in tests.