Server machinery benchmarks
Tests the http flow without the network, but including the switch
that determines whether this is a HTTP1 or HTTP2 connection.

Requests are fixed byte arrays rather than the real client, to test
one thing at a time.

This implementation is quite simplistic, so it will only cover the
most basic path, but that's already something.