http2: ignore unexpected DATA frames in state closed
Fixes #3461

Those frames are somewhat expected at least in the case where we just sent a
RST_STREAM frame and the peer might not have yet processed it.

Of course, it might also be an indication of a misbehaving peer but as long
as there's no indication that this can lead to a non-negligible overhead on
our side, we might be fine with this simple solution.