Make 'network' and 'user' sides more explicit
In Http2ServerSpec information flows in 4 directions:
frames from the network to the server, requests from the server
to the user code, responses from the user code to the server, and
frames from the server to the network.

That can get fairly hard to follow at times. I wonder if it helps
when we explicitly put the test methods in 'user' and 'network'
objects?

(if we like this we should also apply it to Http2ClientSpec of
course)