http2: remove reflective access to alpn classes
Fixes #3808

The former use of structural types tried to use reflection on the concrete
type of the receiver. In this case this was an internal JDK type which cannot
be accessed reflectively in JDK16 any more.

The new pattern relies on lazy class loading for accessing a compatible JDK
only after checking for the method explicitly.