http2: never create Default entity
This was supposed to be a small optimization to avoid having to wrap data elements into `Chunks`
when it was unlikely that we expected a trailing header. The heuristic
was whether we have seen a `content-length` header or not. However, with
more of our optimizations applied, gRPC may actually send a
content-length header on the server side which means we miss out on
trailers on the client side.

Let's revert this optimization for now, so we can receive trailing
headers in all cases again.