http2: never create Default entity (#3841)
This was supposed to be a small optimization to avoid having to wrap data elements into `Chunks`
when it was unlikely that we expected a trailing header. The heuristic
was whether we have seen a `content-length` header or not. However, with
more of our optimizations applied, gRPC may actually send a
content-length header on the server side which means we miss out on
trailers on the client side.

Let's revert this optimization for now, so we can receive trailing
headers in all cases again.

Since we cannot report content-length for some message now any more, some test
bits had to be disabled and there's now https://github.com/akka/akka-http/issues/3843
to find a better solution. For now, the gRPC use case is the most important one.