http2: cache `isDebugEnabled` flag which is somewhat expensive to calculate otherwise
For Slf4j `isDebugEnabled` calls the Slf4jLoggingFilter which needs to look up
the concrete logger with `SLFLoggerFactory.getLogger` for each invocation.

In HTTP/2 we log a lot with debug so caching the value here can make a difference.

The disadvantage is that a dynamic logger level change will not directly apply
to already existing connections but only to new ones which seems like a
useful trade-off.