core: don't include RequestTimeoutSupport if disabled
The documentation of `request-timeout` already hints that you will get
no dynamic request-timeout support if you turn off the feature by setting
it to `infinite`. For some reason never made use of that flexibility.

It turns out that request timeout support is expensive even if it is enabled,
so this will help HTTP/1.1 performance a bit.