core: fail pool slot if connection is canceled before it is failed
Refs #3662

In the issue, it was observed that a connection might only cancel a pool
connection but would not fail it (which was the implicit assumption of the
existing code). In this case a pool might get more and more unresponsive
with slots hanging after connections being canceled but not failed.

The exact condition could not be reproduced but let's be proactive here
and fail the slot if we see unexpected cancellation.