akka-http module Scala 3: Allow to compile without 3.0-migration
This includes explicitly specyfing the type used for scala/java dsl
mapping methods in lambda functions. There appears to be a problem with
the compiler where their type would be incorrectly infered, leading to
the correct JavaMapping not being found and implicit conversions not
occuring.

Also the ambigousness (as reported by the 3.0-migration
option) of references in RoutingSettings was fixed.

Source 3.0-migration option was removed.