core: move pool buffer into PoolInterface again and fix pool sizing
Fixes #4127

Before, we assumed that the pool would always pull `max-connection`
elements which is only correct during regular operation. However,
if connection attempts fail, pool slots do not pull while embargoed.
When that happens, we might in fact not even pull `max-connection` elements
until the connection has healed.

We now keep again a fixed-size buffer directly in the interface. This means
that we are allowing in more elements than before.